Consistent handling of default sources and patches.
rpmbuild permits both "Source:" and "Patch:" settings for the first
(resp. "0th") object, omitting the "0" quantifier. Also the two macros
"%setup" and "%patch" can refer to their respective "defaults" by
omitting the "0". The "%patch" macro may be invoked for a particular
object either as "%patch<number>" or as "%patch -P <number>".

This patch tries to clean up both the setting in the "preample" and the
application in the "%prep" section. Source and patch files are stripped
with the help of the File::Basename->basename() function.