Try to deal with uninstalled 'pax'.
In case that 'pax' is not installed, but unintentionally invoked by
calling 'debbuild' with options '-[bt][as]' or '-i', the output log
should not spill out incorrect text (nor should an empty '.sdeb' file be
left in SDEBS). By invoking $paxcmd in a sub-shell, 'pax' is the final
command and 'system()' can be checked for success. Option '-f' replaces
the redirection to/from STDOUT, so no spurious files will be created.