Rework the logic of %patch processing.
Actually use the first object from 'split /\s+/'. This is the '%patch'
part and may or may not be adorned with a numeric patch number.

Variable '$pnflag' is redundant and can be eliminated.

Matching '(\d*)' may produce the empty string. In conditionals, '0' must
be treated as a valid 'patch number'.

The 'null patch', represented by the empty string '', must only be added
to the patch list, if '%patch' is unadorned and not extended by '-P'.