Permit multiple and combined use of -a/-b options.
This is extreme usage of %setup. Although not really well-documented,
rpm permits cryptic %setup lines like

   %setup -q -T -b 0 -a 1 -a 2

which (0: '-T') suppresses extracting the default 'Source0:',
(1: '-b') extracts 'Source0:' (creating '$name-$version' directory) and
cd's into the builddir, (2: '-a') extracts 'Source1:' and 'Source2:'
(both are expected to be extracted inside the builddir).

By collecting all numeric parameters of the '-a' and '-b' options in
separate lists, it is possible to extend debbuild in a similar way.