Consistent treatment of globals.
Define, set, and modify global variables '_topdir', '_default_patch_flags',
and '_default_patch_fuzz' as elements of '%specglobals'. This permits the
setting of these values in ~/.debmacros and override on the commandline.

rpm made '_default_patch_flags' configurable in 2008 and turned the former
default (-s) off in 2012. Define '%_default_patch_flags -s' in '~/.rpmmacros'
to restore the classic behavior.

In ~/.debmacros, one might be tempted to define

   %_topdir %(echo $HOME)/Work/rpm

so '%(...)' is treated as a 'shell escape'. At the time of writing, this
is just a special case, not a fullblown macro processor.

It is legitimate to use

   rpmbuild -D "_topdir $HOME/path/not/in/rpmmacros" ...

and the same is now possible with debbuild.

It is advertised by the inline documentation that the long options
'--buildroot' and '--define' take their argument after an equal sign.
However, the implementation obviously assumes that the '=' is in fact a
whitespace character, so option and argument are _two_ separate elements
in @ARGV, which are processed separately.

In order to correctly deal with

   debbuild --define="_topdir $HOME/other/than/configured" ...

we process this 'single long option' directly. Not really nice, because
of the code duplication, but at least it works. :o)

Note that this is "just a hack". A fullblown sibling of rpmbuild's
behavior with respect to ~/.rpmmacros and --define/-D would require a
much more thorough investigation and extended implementation.