Yet another form of executive override.
It is legitimate to use

   rpmbuild -D "_topdir $HOME/path/not/in/rpmmacros" ...

and the same is now possible with debbuild.

It is advertised by the inline documentation that the long options
'--buildroot' and '--define' take their argument after an equal sign.
However, the implementation obviously assumes that the '=' is in fact a
whitespace character, so option and argument are _two_ separate elements
in @ARGV, which are processed separately.

In order to correctly deal with

   debbuild --define="_topdir $HOME/other/than/configured" ...

we process this 'single long option' directly. Not really nice, because
of the code duplication, but at least it works. :o)

Note that this is "just a hack". A fullblown sibling of rpmbuild's
behavior with respect to ~/.rpmmacros and --define/-D would require a
much more thorough investigation and extended implementation.