Unified file decompression.
There's an awful lot of code redundancy where we try to find and apply
the correct decompressor for various file types.

However, indepth analysis shows that all decompressors are equal:

   * 'zcat' is the same as 'gunzip -c' is the same as 'gzip -dc'.
   * 'bzcat' is the same as 'bzip2 -dc'.
   * 'xzcat' is the same as 'xz -dc'.

All instances were invoked either as '[bx]zcat' directly or with the
'-dc' options, so there is in fact no variation at all and the need for
ubiquitous redundancy evaporates completely.