All decompressors are equal.
'zcat' is the same as 'gunzip -c' is the same as 'gzip -dc'.
'bzcat' is the same as 'bzip2 -dc'.
'xzcat' is the same as 'xz -dc'.

All instances were invoked either as '[bx]zcat' directly or with
the '-dc' options, so there is no variation at all and the need
for ubiquitous redundancy evaporates.