Fix 'buildroot' handling.
debbuild almost always did _not_ do "the right thing". ;o)

Especially, 'scriptletbase' was expanded too early to catch possible
overrides in the .specfile or on the command-line.  Moreover, '_topdir'
is permitted to be defined with a '%macro' itself, so 'buildroot' might
have to be expanded plusieurs fois.

Note: Not every version of RPM supports the 'BuildRoot:' tag in .spec
files.  Independent of the RPM version, use of the '--buildroot' option
is possible; for details see
'https://fedoraproject.org/wiki/Packaging:Guidelines#BuildRoot_tag'.