Fix --buildroot option.
'BuildRoot' is first defined in 'glomacros' (/usr/lib/debbuild/macros).
Then it can be overridden with the '--buildroot' option. Finally the
specfile is digested, where the 'BuildRoot:' tag takes effect.

We have to intermediately store the '--buildroot' setting to make the
override possible.

Note that some versions of RPM no longer support the  'BuildRoot:' tag.