[RFE #11] debbuild --eval "%macro".
If %macro is defined, its fully expanded definition is printed. Else the
input request '%macro' is reflected.

Note that it is _not_ possible to ask for "%buildroot". Although this is
in fact a macro, it is expanded for internal purposes, so its name is no
longer available in the macro repository after start-up.