Replace deep recursion with linear loop.
By adding the 'g' flag to the regex in the final loop, this steps
through the '$_' input and treats each included '%macro' individually.
Note that this construct can only catch matches in '$1', so '$key' must
be set inside the loop body.

Eliminating the ultimate 'else' leaves undefined '%macros' as-is in the
resulting expansion. Also, the danger of deep-recursion is avoided, so
the nest-level can be removed as well.

The formerly unreachable '%{buildroot}' can now be evaluated.

Example: debbuild --eval "%make_install"