Fix Issue #14 with extreme conditional.
Even with Text::Balanced it is absolutely non-trivial to grok input like

    %{?make_build:%make_build}%{?!make_build:make %{?_smp_mflags}}

Perl requires that all '?' quantifiers on the left of a substitution are
escaped, so we have to neutralize the complete input match.

The input sequence is deleted (i.e., overwritten with the empty string),
when (a) the $macro is defined and negated, or (b) the $macro is not
defined and not negated, so '$macro is defined' and 'negated' have to be
'equivalent'. This can be expressed by 'Not XOR' ('not ... xor ...').