Permit whitespace in regexes.
Learned something new today: Use 'git bisect' to find a bug that got
introduced somewhere in the development history. In case of issue #18,
I introduced an error with the 'Group:' tag between releases 0.15.11
(good) and 15.12.0 (bad); this tag (resp. its 'Section:' counterpart)
remained empty in the resulting 'control' file.

The culprit is 'commit 2d9db3' made Dec 02, 2015, where I tried to
'format' the group (sic!) of preamble tags in the combined regex (and
failed). Second lesson of the day: You have to use the 'x' modifier to
permit formatting whitespace in regexes.