Issue #19: Check for pre-installation dependencies.
'Pre-Depends' is just like 'Depends', except that the listed packages
must be installed, before the current package can even be extracted.

RPM uses the syntax 'Requires(pre)', which we do not support at this
time. (The former 'PreReq' has long been deprecated.)

You are advised to use '%if %{_vendor} == "debbuild"', if you want to
support both syntaxes in the same specfile.