Avoid 'too many arguments' error.
'*glob' entries in the '%files' section more often than not expand to
several items. The '[-d/-f ...]' type checks of the shell can - by their
very nature - work on single items only. So we must not put '*' entries
into the '[...]' test arena. At the time of writing, we will assume a
flat directory, whose 'file' items will be treated as a bunch.

Note: The 'postinst' script is created in 'parse_spec()', at a time when
neither the SOURCE drop is extracted to the 'BUILD' arena, nor the
resulting set of files are installed in the 'BUILDROOT' arena, so we do
not have a safe method to determine the 'type' of an entry at build time.