Simplified '%__scm_apply_quilt' macro.
After studying the source code of 'quilt import', I learned that it is
well able to deal with (most of) our patch files. Despite being called
in a pipe, '%__scm_apply_quilt' simply ignores STDIN and only groks its
input from '%{1}', which RPM correctly expands to whatever is shipped
from the 'patches' list.

DEBBUILD's macro mechanism is somewhat different and does not provide
'%{1}' at this point. However, by expressly specifying the full patch
filename from the 'SOURCES' directory, it is well possible to emulate
the original behaviour.

Note that 'quilt' does _not_ understand '.zip'ped patch files, but RPM
does neither, and this is only an extra feature of DEBBUILD and should
not pose a serious limitation.