Central SCM macro '%apply_patch'.
Try to extract the macronized pipe into a genuine macro.

Note 1: This is the traditional (pre-January 2016) version of this
macro. RPM has an extended version using the '%lua:interpreter'.

Note 2: This initial try results in 'deep recursion', when compiling
'debbuild.spec' of all things! This makes this version unfit for general
application!