Replace global macros.
WTF! Don't do that, I don't like that!

Just because of the one and only '%patch' in 'debbuild.spec' is there oh
_so_ much fun with global macros.

'%optflags' _must_ be replaced as '%{optflags}' (so does '%getconfdir').

'%patch' may or may not be defined as 'numberless Patch:', when it has
to be curly-braced and later path-replaced.