Final fix of issue #25.
Former versions of 'debbuild' did package versioning only for Perl
packages. More recent versions internally work with 'version 0' as
general default; the version stripping occurs only at shipout time.

For C++ programmes, which depend on the 'libstdc++6' library package, it
is very important to accept the '+' character in package names. We use
the general '\S' ('not whitespace') just as in 'getreqs()'.