Debugging framework.
That's absolutely sub-optimal, but what else can I do?
Well, I _could_ start to dig into the original code of
the macro processing in RPM.

Logging output of a successful run.

And now dig into this long protocol and look for '%apply_patch" and its
step-by-step expansion.

The first weird thing I noticed is the duplicate processing of almost
all '%{parametrized macros}' at the beginning and at the end of
'expandmacros()', because of the "special case" '%__scm_apply_SCM'.

The second thing is that '%{expand:...}' is kept far too long in place,
altough '%{elements}' in its replacement are already (pre-)processed.

All in all, the sequence of 'while ( ... bracelet() )' loops is far too
complicated, involved, and convoluted to be "correct" in all cases.