Concise LSB info analysis.
My apologies to those who get in trouble with this. ;o)

I am starting with a new approach to set up the distribution and version
information. This "works for me" on (K)Ubuntu 16.04 LTS "Xenial", but
will most surely break on other variants of Ubuntu/Debian/Linux.

So, please, call back if you have problems with this new version and
tell me what your system provides in order to fill the four entries

   $specglobals{debdist}
   $specglobals{debver}
   $specglobals{$lsb{distributor}}
   $specglobals{dist}

in the %specglobals table.