Rudimentary support for '%post -p <command>'.
'rpm' handles this argument quite differently, i.e., it invokes the
single (!) command without a shell. Plus, it has far more advanced
usages for this syntax, e.g., '-p <lua>', which we most certainly will
never be able to support.

We have to add this part to the respective '%prepost' shell script
anyway, so at least the canonical '%post -p /sbin/ldconfig' should be
working as in 'rpmbuild',