Rework %macro expansion.
The major goal is to permit '%_unmasked_macros', e.g., '%__cp', in the
script sections of our specfiles. This is legal input for 'rpmbuild'.

In order to achieve this, we have to scratch at more than just one spot
in the 'debbuild' code. So bad!

First, for multi-line arguments of 'expandmacros()' that most likely
contain several '%macros', we have to insert a 'while'-loop (throwing
out the superflous 'm'-modifier of the 'm'-matcher).

Second, while scanning single input lines (horribly spread-out code!),
we must only wrap whole multi-word lines in curly braces for '%macros()'
with options. All other '%simplemacros' are treated in 'expandmacros()'.

Next, for my infamous 'debbuild.spec' with the single '%patch' word in
the '%description' section, we have to remove the invocation of
'expandmacros()' for the 'control' contents.

Finally, we notice that the first 'global %macro matcher' is never
actually reached in any of my test cases.

This single commit was created completely 'on the fly', i.e., without
any intermediate Git commits. Don't do this at home, kids!

Many thanks to "Steve Vai and the Wild Things" for background noise!