Add Debian-specific "Breaks" support
The Debian-specific "Breaks" statement is a complement for "Replaces"
when it comes to upgrades. It's a weaker form of "Conflicts" that applies
primarily for apt to select install/update candidates.

This should hopefully not be needed in most circumstances, but in the cases
that it is needed, it is available.

The code is also slightly tweaked to have the appropriate warnings for
all Debian-specific tags not wrapped in a conditional now.