Correctly abort stage scripts and the build when errors occur
In the event that a stage script (that is, the scripts for %prep,
%build, %install, %check, or %clean) ends with a command that exits
with a non-zero error code, the script should fail and the build
should fail. Prior to this change, debbuild did not do this, which
allowed for package builds to complete successfully despite failures
that occurred throughout each stage of the package build.

This change now mimics how rpmbuild(8) calls stage scripts so that
package build failures occur in a similar manner.