macros: Set %_localstatedir and %_sharedstatedir correctly
On virtually all Linux distributions (including Debian and RPM
based ones), %_localstatedir is set to /var and %_sharedstatedir
is set to /var/lib.

In RPM, these are set in platform.in based on what is passed into
the configure script at build time, but debbuild doesn't work that
way.

So, we set it correctly in the macros file so that everyone will
Do the Right Thing(TM).