Add --nodeps support to debbuild
This allows debbuild to be used on non-Debian-based systems to build
Debian packages with the caveat that all dependencies must be fully
specified (since the dependency generator will obviously not work).

This is useful for supporting foreign environments for package builds,
as well as doing cross-target package builds, given the right
environment settings.