Add CI
Add a Circle config to automatically install dependencies and run tests.

Debbuild hasn't formally defined what OS versions or perl versions it
supports. It is currently available in EPEL for CentOS 6. However, use
CentOS 7 because cpanminus (which understands our cpanfile) is easy to
bootstrap; available as an RPM in the default repository set.