Replace spec install section with make install
Move the mechanics of installing out into a Makefile.in. There is
deliberately no Makefile.ac because we're not actually using autotools.

Provide a "configure" script that implements a minimal compatabile
subset of autotools configure functionality, namely, the part where
it composes paths and then substitutes values into a Makefile.in to
create a Makefile. This will be expanded out with more functionality
soon.