Add make build target
A new "build" target to the Makefile to move the build steps out of the
specfile

Slightly alter the way the build works; rather than substituting VERSION
into the same "debbuild" file, produce a new "debbuild.out" that will
then be installed in the same location. This allows "make" to handle
timestamps correctly and reduces the chance that a developer using the
make targets during development will commit the substituted file.