Add make check target
Add new target for running tests. Switch CI to run configure, then a
build, then use the new target to run tests.

Switch the version test to running the build debbuild.out. This
unfortunately means a build is necessary to run tests, but we were going
in that direction anyway, and it will be necessary for testing future
functionality. The test also no longer checks for a specific version,
because that will change with the specfile.