Do not substitute NAME or VERSION in configure
Simplify things by hardcoding the version in the three places it
appears: the script itself, the spec, and the Makefile.in which uses it
to generate documentation. We'll eventually find a more elegant way of
doing this, but this is better than sed-ding into a Perl script and
resolves the weirdness of passing these values to a configure script.

Switch name macros to hardcoding the name, since we never expect this to
change.