Configure version at build-time without sed
Add new mechanism for providing build-time configuration to debbuild.
The old mechanism of sed-ing into the source code was brittle and hard
to get right.

Instead, use the __DATA__ section at the end of the debbuild file to
append an rpmrc-format file with build-time configuration. debbuild
parses this at startup and uses it if it exists. If not, it falls back
to default values to ensure that debbuild will still run directly in its
source form.

This mechanism is initially being used for the version but will
eventually be extended to provide other build-time configuration.