Handle multiple debrc entries per name
Fix a regression where only the last optflags entry in the debrc was
preserved.

When debbuild switched to using read_debrc_from_handle, a bug was
introduced where a subsequent entry with a duplicate name overwrote
previous ones, since the return value of that sub was stored as a simple
hash. (I didn't realize at the time that the rpmrc format allowed
multiple entries with the same name.)

Enhance read_debrc_from_handle to return a hash of arrayrefs, containing
each of the entires for a particular name.