Load debrc and macros during tests
Allow test runs that use a debbuild built to reference files from this
repository rather than their installed locations. This will enable
complete end-to-end automated tests that use macros and debrc without
installing them.

During "make check", set a DESTDIR of $pwd/makecheck and perform a "make
install" into that directory. Build a "debbuild.out" with paths that
point to this directory. Then, invoke the makecheck/bin/debbuild in
tests.

This has a substantial caveat that `debbuild.out` clashes between "make
build" and "make check", that is, that "make" isn't smart enough to
rebuild the file if one target is invoked after the other. Suggestions
on how to deal with this are welcome.

Introduce a new SCRIPTPATHPREFIX make variable that's normally an empty
string, but set to the DESTDIR during "make check". We never want
DESTDIR to be part of paths baked into debbuild because DESTDIR is
usually a temporary directory made during package build that doesn't
reflect the final installed location. However, we're using DESTDIR in
this case to redirectt files to the "makecheck" directory, so we need a
mechanism to make debbuild look there.