Remove unnecessary escaping for Requires(pre)/Requires(preun)
With 51dccbd392f7d82ea1d30f53d86e811ccfb11994, debbuild switched
from using regular expressions to do fuzzy matching on stanzas
declared in a subpackage to using exact literal matches.

However, the extra escaping of parens that was supposed to be removed
as part of changing to using exact literal matching was left in,
causing the keys read in for 'Requires(pre)' and 'Requires(preun)'
to be escaped and thus fail to match with the right value in the
quote word group.

This extra escaping was necessary with the fuzzy matching logic
because otherwise the parens would be considered as a capturing
group in a regular expression. Since we do not use regular
expressions to identify the stanzas anymore, this became
unnecessary.

Dropping the statement that does the extra escaping fixes the
handling for these dependency stanzas.