Add support for gems as source files
Currently, it's not possible to use a gem as a source file like in
other tools such as rpmbuild. This means that the source can't be
obtained from rubygems.org, which is often desirable, and separate
deb-specific logic is required when packaging gems for both deb and
rpm-based systems. This adds support for using .gem files directly.