Fix detection of subpackage naming format
In circumstances where a subpackage name had '-n' in it, debbuild
would incorrectly match on that and fail to put the expected prefix
for binary package names declared using the subpackage shorthand
form of "%package bar" such that "%package bar-nox" for source package
"foo" would make a package "bar-nox" instead of "foo-bar-nox".

Fix this by ensuring that we match on '-n' used as an option by
checking also for a whitespace character immediately before and after,
as that is obviously not a package name if whitespace characters exist.