expandmacros: replace gobal macros one at time
replacemacros() is looping all available macros in input, to convert
them to %{} syntax. The conversion used to use s///g, which causes all
instances to be replaced in one pass. This causes wrong replaces in some
cases. Consider following input:

%cmake
%cmake_build

When one replaces %cmake with %{cmake} with g flag, latter line would be
%{cmake}_build, causing wrong expansion to happen at later point.