Improve README, add bootstrap script for self hosted installation
For new users who haven't set up debbuild before, it's not immediately
obvious how to install it as a package. This is annoying, and has
frustrated me (and I'm sure many others) many times over the years,
especially when preparing to bootstrap a new Debian or Ubuntu release.

This pull request adds a simple shellscript, `bootstrap.sh`, which builds
a `.deb` of debbuild, and tells the user where it is so they can install
it. Reference is made to this script in the README, and basic usage
instructions have also been added.

Signed-Off-By: Dan Fuhry <dan@fuhry.com>