Fix %include handling by passing file handle
I have a project that uses %include and it fails to build with:

readline() on closed filehandle $fh at /usr/bin/debbuild line 607.

In fact, the test case in testdir/dummy.spec will also demonstrate this
behavior.

This is due to commit 3d0dba6e (Handle multiline macros in specfiles)
turning $fh into a global.  When parse_spec opens the included file, it
reuses the global file handle variable and when it completes, it closes it,
leading to failure.

This commit fixes the issue by restoring the behavior of parse_spec
owning its file handle and then passing it as an argument to consumers.
Due perl's open operation only creating new file handles when the
variable to be used is undefined, it's insufficient to keep a copy of
the variable when parse_spec recurses.