Use %_debbuild instead of vendor for debbuild conditionals
Distros are meant to set the vendor, not the tool. So if debbuild
gets included in a distribution it may have a different vendor
than 'debbuild'. Rpm doesn't have rpm as vendor either after all.
Debian for example defines _vendor as "debian" for rpmbuild. The
internal _debbuild macro is set to the version of debbuild so seems
like a good way to identify the tool