Avoid empty calls of dpkg-query for libc.so.6.
This PR improves on PR #255 on (K)Ubuntu and possibly Debian systems.

For whatever reason, dpkg-query does not seem to have an option to
follow symlinked paths on its own. System stuff like libc.so.6 is
located under path /usr/lib, which is symlinked to path /lib.

By replacing the absolute path of such library files, we can skip the
double invocation of dpkg_query_paths and the quite involved resolution
to the same result (including a call to abs_path at that).

This PR was developed on Kubuntu 24.04.1 LTS with debbuild 24.12.0 with
example package <https://github.com/ascherer/mmix/blob/local/mmix.spec>.