Issue #265: Purge old line numbers in DEBBUILD.POT.
By leaving out the '-j' option of the 'xgettext' incantation and by
relying on 'git diff' to restore vital header information, we can easily
create an up-to-date version of the DEBBUILD.POT file that truely
matches the DEBBUILD program. This eliminates not only shifted/outdated
line information for the translatable strings but also several strings
that are no longer in use.